Load best rates from the SQLDelight

+++ data-testing/src/commonMain/kotlin/dev/fobo66/core/data/testing/fake/FakeFormattingDataSource.kt
@@ -20,7 +20,7 @@ import fobo66.valiutchik.core.entities.LanguageTag
 import fobo66.valiutchik.core.model.datasource.FormattingDataSource
 
 class FakeFormattingDataSource : FormattingDataSource {
-    override fun formatCurrencyValue(value: Float, languageTag: LanguageTag): String =
+    override fun formatCurrencyValue(value: Double, languageTag: LanguageTag): String =
         value.toString()
 
     override fun formatBankName(name: String, languageTag: LanguageTag): String = name
